--- a/composer.json
+++ b/composer.json
@@ -9,7 +9,7 @@
     "xp-framework/core": "^7.0 | ^6.5",
     "xp-framework/http": "^7.0 | ^6.2",
     "xp-framework/xml": "^7.0 | ^6.0",
-    "xp-framework/scriptlet": "^8.0.1 | ^6.2",
+    "xp-framework/scriptlet": "^8.0.1",
     "xp-framework/logging": "^7.0 | ^6.5",
     "xp-framework/collections": "^7.0 | ^6.5",
     "xp-framework/networking": "^7.0 | ^6.6",