--- a/src/test/php/webservices/rest/unittest/RestJsonSerializerTest.class.php
+++ b/src/test/php/webservices/rest/unittest/RestJsonSerializerTest.class.php
@@ -35,14 +35,14 @@ class RestJsonSerializerTest extends TestCase {
     $this->assertEquals('"'.$str.'"', $this->serialize($str));
   }
 
-  #[@test, @values([-1, 0, 1, 4711])]
-  public function integers($int) {
-    $this->assertEquals(''.$int, $this->serialize($int));
+  #[@test, @values([['-1', -1], ['0', 0], ['1', 1], ['4711', 4711]])]
+  public function integers($expected, $int) {
+    $this->assertEquals($expected, $this->serialize($int));
   }
 
-  #[@test, @values([-1.0, 0.0, 1.0, 47.11])]
-  public function decimals($decimal) {
-    $this->assertEquals(''.$decimal, $this->serialize($decimal));
+  #[@test, @values([['-1.0', -1.0], ['0.0', 0.0], ['1.0', 1.0], ['47.11', 47.11]])]
+  public function decimals($expected, $decimal) {
+    $this->assertEquals($expected, $this->serialize($decimal));
   }
 
   #[@test]
@@ -57,23 +57,23 @@ class RestJsonSerializerTest extends TestCase {
 
   #[@test]
   public function empty_array() {
-    $this->assertEquals('[ ]', $this->serialize([]));
+    $this->assertEquals('[]', $this->serialize([]));
   }
 
   #[@test]
   public function int_array() {
-    $this->assertEquals('[ 1 , 2 , 3 ]', $this->serialize([1, 2, 3]));
+    $this->assertEquals('[1,2,3]', $this->serialize([1, 2, 3]));
   }
 
   #[@test]
   public function string_array() {
-    $this->assertEquals('[ "a" , "b" , "c" ]', $this->serialize(['a', 'b', 'c']));
+    $this->assertEquals('["a","b","c"]', $this->serialize(['a', 'b', 'c']));
   }
 
   #[@test]
   public function string_map() {
     $this->assertEquals(
-      '{ "a" : "One" , "b" : "Two" , "c" : "Three" }',
+      '{"a":"One","b":"Two","c":"Three"}',
       $this->serialize(['a' => 'One', 'b' => 'Two', 'c' => 'Three'])
     );
   }
@@ -84,7 +84,7 @@ class RestJsonSerializerTest extends TestCase {
   #])]
   public function traversable_array($in) {
     $this->assertEquals(
-      '[ 1 , 2 , 3 ]',
+      '[1,2,3]',
       $this->serialize($in)
     );
   }
@@ -95,7 +95,7 @@ class RestJsonSerializerTest extends TestCase {
   #])]
   public function traversable_map($in) {
     $this->assertEquals(
-      '{ "color" : "green" , "price" : "$12.99" }',
+      '{"color":"green","price":"$12.99"}',
       $this->serialize($in)
     );
   }
@@ -106,7 +106,7 @@ class RestJsonSerializerTest extends TestCase {
   #])]
   public function empty_traversable($in) {
     $this->assertEquals(
-      '[ ]',
+      '[]',
       $this->serialize($in)
     );
   }