## Performance
Using slightly less memory, we get a 7-fold permance increase.

Dimension | Before | After |
| ----------- | ----------- | --------- |
| Runtime | 10.833 seconds | 1.481 seconds |
| Iteration | 108.3 ms per iteration, result = 30 | 14.8 ms per iteration, result = 30 |
| Memory | 1176.195 kB memory used | 1055.477 kB memory used |
| Peak memory | 1573.680 kB peak | 1101.750 kB peak |

*Parsing a 150 kB JSON file*

## Fixes

* The JSON library handles surrogate pairs correctly - see https://github.com/xp-forge/json/pull/6
* The JSON library also handles nested [traversables](http://php.net/traversable) correctly

## Possible BC break

We need a fix in the scriptlet library and cannot have compat with ^6.2, ^7.0 and 8.0 but require [the 8.0.1 bugfix release](https://github.com/xp-framework/scriptlet/releases/tag/v8.0.1) to work correctly. The scriptlet library does not introduce any problematic changes for 7.0 and 8.0 but rather chose to announce major features with these versions, so it might not be relevant in the real word!

## JSON output changes

The output format is now more dense by default - there are no superfluous spaces or newlines in the JSON output. Also, slashes are no longer escaped (e.g. `http:\/\/example.com\/`). Tests relying on this exact behavior will need to be changed.

## Versioning

Given the changes above - most probably won't break any real-life situation but could cause trouble if people have relied on the whitespace in JSON I suggest to release this as a major version (8.0.0) so people stop, think, and maybe read this before upgrading /cc @lluchs @kiesel