@thekid 2016-06-08
Talked with @kiesel and decided 7.2.0 would be OK for this, too.