@thekid 2016-06-08
Typical update scenario looks like this:

```sh
friebe@vagrant-friebe-jessie64:~/dev/bitbucket-sync$ composer up
Loading composer repositories with package information
Updating dependencies (including require-dev)
  - Removing xp-framework/webservices (v7.0.0)
  - Removing xp-framework/parser (v7.0.0)
  - Installing xp-forge/json (v2.1.0)
    Downloading: 100%

  - Removing xp-framework/rest (v7.1.0)
  - Installing xp-framework/rest (v7.2.0)
    Downloading: 100%

Writing lock file
```