@thekid 2016-06-08
:shipit: Released as [7.2.0](https://github.com/xp-framework/rest/releases/tag/v7.2.0)